+++ imswitch/imcommon/model/__init__.py
@@ -3,4 +3,4 @@ from .VFileCollection import VFileItem, VFileCollection
 from .api import APIExport, generateAPI
 from .logging import initLogger
 from .shortcut import shortcut, generateShortcuts
-from .utils import angleToSteps, stepsToAngle
+from .mathutils import angleToSteps, stepsToAngle
