+++ imswitch/imcommon/model/mathutils.py
